requirements.txt
fastapi~=0.116.1
uvicorn~=0.35.0
sqlalchemy~=2.0.41
python-dotenv~=1.1.0
psycopg2-binary
pydantic[email]~=2.11.7
httpx~=0.28.1
pytest~=8.4.1
starlette~=0.46.2
aio-pika~=9.5.5
requests~=2.31.0
psycopg~=3.2.9